fixed a few more pages to adhear to the new layout
